Create custom tax calculator to apply zip code zone rates only
Copy over the default tax calculator to a custom one to allow
updating the logic that gets list of zones to use to calculate
tax rate.

If a zip code zone is present, then only that will be returned. Only
when none will be found will we continue to applying rates for
state/country.